+++ src/apps/coreutils/readelf/readelf.cc
@@ -47,7 +47,7 @@ static void parse(FStream &bin) {
     cout << "Program Headers:\n";
     cout << "  Type    Offset     VirtAddr     PhysAddr     FileSiz   MemSiz    Flg Align\n";
 
-    off_t off = header.e_phoff;
+    size_t off = header.e_phoff;
     for(uint i = 0; i < header.e_phnum; ++i, off += header.e_phentsize) {
         ELF_PH pheader;
         if(bin.seek(off, SEEK_SET) != off)
@@ -66,7 +66,7 @@ static void parse(FStream &bin) {
              << ((pheader.p_flags & PF_X) ? "E" : " ") << " "
              << fmt(pheader.p_align, "#0x") << "\n";
 
-        if(bin.seek(pheader.p_offset, SEEK_SET) != (off_t)pheader.p_offset)
+        if(bin.seek(pheader.p_offset, SEEK_SET) != pheader.p_offset)
             exitmsg("Invalid ELF-file");
 
         size_t count = pheader.p_filesz;
