+++ src/apps/pipe/pipe.cc
@@ -40,7 +40,7 @@ int main() {
 
         reader.run([] {
             File *in = VPE::self().fds()->get(STDIN_FD);
-            size_t res;
+            ssize_t res;
             while((res = in->read(buffer, sizeof(buffer))) > 0)
                 cout << "[1] Read " << res << ": '" << buffer << "'\n";
             return 0;
@@ -53,7 +53,7 @@ int main() {
 
         reader2.run([] {
             File *in = VPE::self().fds()->get(STDIN_FD);
-            size_t res;
+            ssize_t res;
             while((res = in->read(buffer, sizeof(buffer))) > 0)
                 cout << "[2] Read " << res << ": '" << buffer << "'\n";
             return 0;
@@ -103,7 +103,7 @@ int main() {
 
         reader.run([] {
             File *in = VPE::self().fds()->get(STDIN_FD);
-            size_t res, i = 0;
+            ssize_t res, i = 0;
             while(i++ < 3 && (res = in->read(buffer, sizeof(buffer))) > 0)
                 cout << "Read " << res << ": '" << buffer << "'\n";
             return 0;
@@ -174,7 +174,7 @@ int main() {
         pipe.close_writer();
 
         File *in = VPE::self().fds()->get(pipe.reader_fd());
-        size_t res, i = 0;
+        ssize_t res, i = 0;
         while(i++ < 3 && (res = in->read(buffer, sizeof(buffer))) > 0)
             cout << "Read " << res << ": '" << buffer << "'\n";
 
@@ -193,7 +193,7 @@ int main() {
 
         reader.run([] {
             File *in = VPE::self().fds()->get(STDIN_FD);
-            size_t res, i = 0;
+            ssize_t res, i = 0;
             while(i++ < 3 && (res = in->read(buffer, sizeof(buffer))) > 0)
                 cout << "Read " << res << ": '" << buffer << "'\n";
             return 0;
@@ -251,9 +251,9 @@ int main() {
         t2.run([] {
             File *in = VPE::self().fds()->get(STDIN_FD);
             File *out = VPE::self().fds()->get(STDOUT_FD);
-            size_t res;
+            ssize_t res;
             while((res = in->read(buffer, sizeof(buffer))) > 0) {
-                out->write(buffer, res);
+                out->write(buffer, static_cast<size_t>(res));
             }
             return 0;
         });
@@ -264,7 +264,7 @@ int main() {
         p2.close_writer();
 
         File *in = VPE::self().fds()->get(p2.reader_fd());
-        size_t res;
+        ssize_t res;
         while((res = in->read(buffer, sizeof(buffer))) > 0)
             cout << "Received " << res << "b: '" << buffer << "'\n";
